Bug fix of Map prop remove

+++ app/src/game/world.h
@@ -6,14 +6,14 @@
 
 bool world_system_initialize(camera_metrics* _in_camera_metrics);
 
-void set_worldmap_location(u16 id);
+void set_worldmap_location(i32 id);
 worldmap_stage* get_worldmap_locations(void);
 worldmap_stage* get_active_worldmap(void);
 void set_map_tile(i32 layer, tile src, tile dst);
 tilemap* get_active_map(void);
 tilemap_prop_address get_map_prop_by_pos(Vector2 pos);
-tilemap_prop_static* get_map_prop_static_by_id(u16 id);
-tilemap_prop_sprite* get_map_prop_sprite_by_id(u16 id);
+tilemap_prop_static* get_map_prop_static_by_id(i32 id);
+tilemap_prop_sprite* get_map_prop_sprite_by_id(i32 id);
 
 void save_current_map(void);
 void load_current_map(void);
@@ -23,15 +23,15 @@ tile _get_tile_from_map_by_mouse_pos(u16 from_layer, Vector2 _mouse_pos);
 
 bool add_prop_curr_map(tilemap_prop_static prop_static);
 bool add_prop_curr_map(tilemap_prop_sprite prop_sprite);
-bool remove_prop_cur_map_by_id(u16 id, tilemap_prop_types type);
+bool remove_prop_cur_map_by_id(i32 id, tilemap_prop_types type);
 void update_map(void);
 void drag_tilesheet(Vector2 vec);
 void _render_tile_on_pos(tile* _tile, Vector2 pos, tilesheet* sheet);
 void render_map(void);
 void render_map_view_on(Vector2 pos, f32 zoom);
 void render_map_palette(f32 zoom);
-bool change_prop_zindex(tilemap_prop_types type, u32 id, i16 old_zindex, i16 new_zindex);
-void refresh_render_queue(u16 id);
+bool change_prop_zindex(tilemap_prop_types type, i32 id, i16 old_zindex, i16 new_zindex);
+void refresh_render_queue(i32 id);
 
 #define _remove_prop_cur_map_by_id(PROP) remove_prop_cur_map_by_id(PROP->id, PROP->prop_type)
